+++ src/interfaces/bridge/ILeafRestrictedTokenBridge.sol
@@ -6,7 +6,7 @@ import {ITokenBridge} from "./ITokenBridge.sol";
 interface ILeafRestrictedTokenBridge is ITokenBridge {
     /// @notice Chain id for Base network
     /// @return The Base chain id (8453)
-    function BASE_CHAIN_ID() external view returns (uint32);
+    function BASE_CHAIN_ID() external view returns (uint256);
 
     /// @notice Address of the voter contract
     /// @return The address of the voter contract
